fix: SwiftData persistence — explicit saves, schema versioning, pre-deploy validation
Data was silently lost because modelContext.insert() was never followed
by an explicit save(), and no VersionedSchema existed to handle schema
changes between builds. Now all models live in RoboSchema.swift as the
single source of truth with migration plan infrastructure.

Also adds attribution links, dynamic build number, and validate-build.sh.
