+++ ios/project.yml
@@ -12,7 +12,7 @@ settings:
     CODE_SIGN_STYLE: Automatic
     PRODUCT_NAME: Robo
     MARKETING_VERSION: 1.0
-    CURRENT_PROJECT_VERSION: 5
+    CURRENT_PROJECT_VERSION: 6
     SWIFT_VERSION: 5.9
     IPHONEOS_DEPLOYMENT_TARGET: 17.0
     TARGETED_DEVICE_FAMILY: 1  # iPhone only
